Update system to use dbus and update cmd interface
This is needed to the GUI can be run as non-root. That said,
this commit doesn't actually update the GUI. There is a lot of refactoring
but few features outside of dbus and a few bug fixes

Weresync now uses a daemon which must be started manually. Ideally
distribution packages will include a way to run this automatically. An example
service file for systemctl is included in the resources folder.

The dbus policy file (weresync/resources/weresync-dbus.conf) must be copied to
/etc/dbus-1/system.d or the daemon will not function. Again packages should
do this automatically, it is outside of pip's realm.